refactor : member_character_book -> member_character 로 변경
