--- src/main/java/univ/earthbreaker/namu/database/core/MemberCharacterBookJpaRepository.java
@@ -1,6 +1,0 @@
-package univ.earthbreaker.namu.database.core;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberCharacterBookJpaRepository extends JpaRepository<MemberCharacterBookJpaEntity, Long> {
-}
