+++ src/main/java/univ/earthbreaker/namu/database/core/character/MemberCharacterJpaEntity.java
@@ -1,4 +1,4 @@
-package univ.earthbreaker.namu.database.core;
+package univ.earthbreaker.namu.database.core.character;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
